requirements.txt
cuda-python
einops
imageio
natsort
opencv-python
git+https://github.com/openai/CLIP.git
pandas
prefetch_generator
pillow
python-dotenv
pytorch_lightning
regex
matplotlib
matlabengine
numpy
scipy
setuptools
scikit-image
smop
torch
torchio
torchsummary
torchvision
tqdm